lots of cleanup
