+++ packages/scatterbrain/src/demo.html
@@ -22,7 +22,7 @@
       place-content: center center;
     }
   </style>
-  <script defer src="../dist/module.js" type="module"></script>
+  <script defer src="../dist/main.js" type="module"></script>
 </head>
 
 <body>
