+++ packages/scatterbrain/src/index.ts
@@ -9,4 +9,3 @@ export {
 export { buildScatterbrainCacheClient } from './cache-client'
 export * from './types';
 export { getVisibleItems, loadDataset as loadScatterbrainDataset } from './dataset';
-export { whatever } from './tgpu-shader'
